twisterlib: remove unintended test cases from test plan
Due to the diverse coding style and lack of preprocessing when
scanning for test cases, there are many unintended test suite/
test function combinations in the initial test plan.
This patch tries to remove those unintended ones by checking the
function exsitence in the elf file. The check is done right
after the build stage of the pipeline.

Signed-off-by: Ming Shao <ming.shao@intel.com>